Add performance module - move app startup measurement there
