+++ settings.gradle
@@ -17,7 +17,7 @@ include ':app',
   ':feature:in-app-update',
   ':feature:config-debug-api',
   ':feature:config-debug',
-
+  ':feature:performance',
   ':feature:users',
   ':feature:repo',
   ':feature:push',
